🏷️Update list types and move NavItem to nav templatetags
